Extract Selection class to its own file
