+++ __init__.py
@@ -15,7 +15,6 @@
 # You should have received a copy of the GNU General Public License
 # along with Japanese Furigana.  If not, see <http://www.gnu.org/licenses/>.
 
-import json
 import re
 import os
 
@@ -24,11 +23,11 @@ from aqt.qt import *
 
 from aqt import mw
 
-from anki.buildinfo import version
 from anki.hooks import addHook
 
 from . import reading
 from . import config
+from .selection import Selection
 
 mecab  = reading.MecabController()
 config = config.Config()
@@ -82,45 +81,6 @@ def deleteFurigana(editor, s):
             tooltip(_("No furigana found to delete"))
         else:
             s.modify(html)
-
-class Selection:
-
-    js_get_html = u"""
-        var selection = window.getSelection();
-        var range = selection.getRangeAt(0);
-        var div = document.createElement('div');
-        div.appendChild(range.cloneContents());
-        div.innerHTML;
-    """
-
-    def __init__(self, window, callback):
-        self.window = window
-        self.setHtml(None, callback)
-
-    def isDeprecated(self):
-        return int(version.replace('.', '')) < 2141
-
-    def setHtml(self, elements, callback, allowEmpty=False):
-        self.selected = elements
-        if self.selected == None:
-            if self.isDeprecated():
-                self.window.web.eval("setFormat('selectAll');")
-                self.window.web.page().runJavaScript(self.js_get_html, lambda x: self.setHtml(x, callback, True))
-            else:
-                self.window.web.page().runJavaScript("getCurrentField().fieldHTML", lambda x: self.setHtml(x, callback, True))
-            return
-        self.selected = self.convertMalformedSpaces(self.selected)
-        callback(self)
-
-    def convertMalformedSpaces(self, text):
-        return re.sub(r'& ?nbsp ?;', ' ', text)
-
-    def modify(self, html):
-        html = self.convertMalformedSpaces(html)
-        if self.isDeprecated():
-            self.window.web.eval("setFormat('insertHTML', %s);" % json.dumps(html))
-        else:
-            self.window.web.page().runJavaScript("getCurrentField().fieldHTML = %s;" % json.dumps(html))
 
 setupGuiMenu()
 addHook("setupEditorButtons", addButtons)
